Refactor CreateRatePage and hooks: integrate useNoveltyTypes for dynamic novelty options, remove obsolete useLocations and useRates hooks, and enhance INoveltyType with service_type_id.
